+++ src/util/data.ts
@@ -1,4 +1,5 @@
 import type {Game} from "@core/game";
+import {HexColor} from "@core/types";
 
 /**
  * @param obj1 source object
@@ -156,9 +157,9 @@ export function safeClone<T>(obj: T): T {
 
 export type Values<T> = T[keyof T];
 
-export function toHex(hex: { r: number; g: number; b: number; a?: number } | string): string {
+export function toHex(hex: { r: number; g: number; b: number; a?: number } | string): HexColor {
     if (typeof hex === "string") {
-        return hex;
+        return hex as HexColor;
     }
     return `#${(hex.r || 0).toString(16).padStart(2, "0")}${(hex.g || 0).toString(16).padStart(2, "0")}${(hex.b || 0).toString(16).padStart(2, "0")}${(hex.a === undefined ? "" : hex.a.toString(16).padStart(2, "0"))}`;
 }